+++ client/.gitignore
@@ -2,3 +2,6 @@ node_modules
 dist
 out
 *.log*
+
+# Sentry Config File
+.env.sentry-build-plugin
