+++ client/electron.vite.config.js
@@ -1,3 +1,4 @@
+import { sentryVitePlugin } from '@sentry/vite-plugin'
 import { resolve } from 'path'
 import { defineConfig, externalizeDepsPlugin } from 'electron-vite'
 import react from '@vitejs/plugin-react'
@@ -6,9 +7,11 @@ export default defineConfig({
   main: {
     plugins: [externalizeDepsPlugin()]
   },
+
   preload: {
     plugins: [externalizeDepsPlugin()]
   },
+
   renderer: {
     resolve: {
       alias: {
@@ -16,5 +19,16 @@ export default defineConfig({
       }
     },
     plugins: [react()]
-  }
+  },
+
+  build: {
+    sourcemap: true
+  },
+
+  plugins: [
+    sentryVitePlugin({
+      org: 'timely-v2',
+      project: 'javascript-react'
+    })
+  ]
 })
