+++ client/package.json
@@ -30,6 +30,9 @@
     "@fullcalendar/react": "^6.1.9",
     "@fullcalendar/timegrid": "^6.1.9",
     "@mui/icons-material": "^5.14.16",
+    "@sentry/react": "^7.81.1",
+    "@sentry/tracing": "^7.81.1",
+    "@sentry/vite-plugin": "^2.10.1",
     "date-fns": "^2.30.0",
     "electron-updater": "^6.1.1",
     "immutability-helper": "^3.1.1",
