--- client/src/renderer/src/components/ErrorBoundary.jsx
@@ -1,53 +1,0 @@
-import { Component } from 'react'
-import PropTypes from 'prop-types'
-
-class ErrorBoundary extends Component {
-  constructor(props) {
-    super(props)
-    this.state = { hasError: false, error: null, errorInfo: null }
-  }
-
-  static getDerivedStateFromError(error) {
-    return { hasError: true }
-  }
-
-  componentDidCatch(error, errorInfo) {
-    this.setState({
-      error: error,
-      errorInfo: errorInfo
-    })
-
-    logErrorToMyService(error, errorInfo)
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return (
-        <div>
-          <h1>Something went wrong.</h1>
-          <details style={{ whiteSpace: 'pre-wrap' }}>
-            {this.state.error && this.state.error.toString()}
-            <br />
-            {this.state.errorInfo.componentStack}
-          </details>
-        </div>
-      )
-    }
-
-    return this.props.children
-  }
-}
-
-ErrorBoundary.propTypes = {
-  children: PropTypes.node.isRequired
-}
-
-export default ErrorBoundary
-
-function logErrorToMyService(error, errorInfo) {
-  console.error('ErrorBoundary caught an error', error, errorInfo)
-
-  // Here you would send the error and errorInfo to your logging service
-  // For example, if you're using a service like Sentry:
-  // Sentry.captureException(error, { extra: errorInfo });
-}
